+++ packages/slate-vue/plugins/vue-editor.ts
@@ -1,13 +1,9 @@
 /**
  * a copy of slate-react
  */
-import {
-  VueEditor
-} from 'slate-vue-shared'
+import { VueEditor } from 'kessel-slate-vue-shared';
 import { vueRuntimeFunc } from './vue-runtime';
 
-VueEditor.toDOMRange = vueRuntimeFunc(VueEditor.toDOMRange)
+VueEditor.toDOMRange = vueRuntimeFunc(VueEditor.toDOMRange);
 
-export {
-  VueEditor
-}
+export { VueEditor };